Refactor creating challenge
