+++ packages/server/src/app/challenge/controllers/user_challenge.controller.js
@@ -1,12 +1,6 @@
 import { UserChallengeService } from '../services/user_challenge.service.js';
 
 class UserChallengeController {
-  async create(userChallengeData) {
-    console.info('UserChallengeController - Creating a new UserChallenge resources');
-
-    return await UserChallengeService.instance().create(userChallengeData);
-  }
-
   /**
    * @param challengeIdParam {string}
    * @return {Promise<UserChallengeEntity[]>}
