+++ packages/server/src/app/group/index.js
@@ -1,0 +1,1 @@
+export * from './services/user_group.service.js';
